ml2_clases/unidad5_practica/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-huggingface>=0.0.3
chromadb>=0.4.0
langchain-chroma>=0.1.0
pypdf>=3.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
sentence-transformers>=2.2.0
